+++ src/main/java/com/koddy/server/member/presentation/dto/request/SignUpRequest.java
@@ -1,7 +1,7 @@
 package com.koddy.server.member.presentation.dto.request;
 
 import com.koddy.server.member.domain.model.MemberType;
-import com.koddy.server.member.utils.ValidMailCheck;
+import com.koddy.server.member.utils.validator.ValidMailCheck;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
@@ -10,7 +10,7 @@ public record SignUpRequest(
         String email,
 
         @ValidMailCheck
-        boolean checked,
+        Boolean checked,
 
         @NotBlank(message = "패스워드는 필수입니다.")
         String password,
